Fix 1 big and 1 small memory leak.
1) When client closed the SSE connection (by closing the tab),
it appears that the connection on server side wasn't closed properly.
Most likely the socket (or something else) was still kept open due to
the "Connection": "keep-alive" header. Everytime you then tried to send
message to the "closed" connection with res.write(), it resulted into
slow but steady memory leak. This lead into the server crashing after
exactly 72 hours after booting. It was fixed by manually closing the
connection on "close" event by calling res.end(). I also added
res.socket.destroy() to be 100% sure that the connection is completely
done.

2) When new SSE connection was estabilished, a recursive loop was
started and it would run every 30s. Point of this loop was
to send keep-alive messages to the client. Problem was that there was
nothing that would ever cancel this loop. That means even after the
connection was closed, the loop would still run every 30s and try to
send the keep-alive message (which was the source of the main memory
leak). But that also means that eventually, if enough people visited
the site over time, there would just be so many loops running at
the same time that it would crash the server too (even without
the main memory leak).